HDR textures loading, scene hierarchy add fixed, skybox for metal boilerplate [wip]

+++ src/Engine/Render/Metal/Shaders/PbrMaterial.metal
@@ -13,19 +13,19 @@ struct VertexOut {
 };
 
 vertex VertexOut pbr_vertex(const device BasicVertexLayout *vertexArray [[buffer(0)]], const device UniformsData &uniformsData [[buffer(1)]], unsigned int vid [[vertex_id]]) {
-  BasicVertexLayout in = vertexArray[vid];
+    BasicVertexLayout in = vertexArray[vid];
 
-  VertexOut out;
+    VertexOut out;
 
-  out.color = in.pos;
-  out.uv = in.uv;
-  out.pos = uniformsData.projection * uniformsData.view * uniformsData.model * in.pos;
+    out.color = in.pos;
+    out.uv = in.uv;
+    out.pos = uniformsData.projection * uniformsData.view * uniformsData.model * in.pos;
 
-  return out;
+    return out;
 }
 
-fragment float4 pbr_fragment(VertexOut interpolated [[stage_in]], texture2d<half, access::sample> colorTexture [[ texture(BaseColorTexture) ]], sampler textureSampler [[ sampler(BaseColorTexture) ]]) {
-    const half4 colorSample = colorTexture.sample(textureSampler, interpolated.uv);
+fragment float4 pbr_fragment(VertexOut interpolated [[stage_in]], texture2d<float, access::sample> colorTexture [[ texture(BaseColorTexture) ]], sampler textureSampler [[ sampler(BaseColorTexture) ]]) {
+    const float4 colorSample = colorTexture.sample(textureSampler, interpolated.uv);
 
     return float4(colorSample.rgba);
 }
